add max limit for ban period

+++ src/main/java/com/example/my_bot/exception/ban/BanPeriodOutOfBoundsException.java
@@ -6,7 +6,7 @@ import static com.example.my_bot.utils.TimeUtils.formatDurationFromSeconds;
 
 public class BanPeriodOutOfBoundsException extends BanException {
 
-    public BanPeriodOutOfBoundsException(int minInSeconds) {
+    public BanPeriodOutOfBoundsException(long minInSeconds) {
         super("Минимальный период для временного бана — %s"
                 .formatted(formatDurationFromSeconds(minInSeconds, true)));
     }
